requirements.txt
asgiref==3.3.4
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.2.2
django-csp==3.7
django-ranged-response==0.2.0
django-ratelimit==3.0.1
django-recaptcha==2.0.6
folium==0.12.1
idna==2.10
Jinja2==3.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.4.2
numpy==1.20.3
paho-mqtt==1.5.1
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
